requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
greenlet==2.0.2
idna==3.4
mysql-connector-python==8.0.33
numpy==1.25.0
pandas==2.0.2
protobuf==3.20.3
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
